Fixing tests
- Removed tests that depend on constant values and API keys that are not allowed to be uploaded to a public git repo.

Ideally a Unit test shouldn't test if another API is working or not, it should only check whether the library is in knows how to deal with the data it returns and possible errors. That's why I'll add more tests in the future which cover these scenarios.

Currently the bot doesn't do much, it only runs the request and returns the response, that's why I'll add more tests in the future, when some more fancy treatments are in.